Allow filtering by stealth level in `pkg list`
